requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
cachetools==5.2.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
google-api-core==2.8.2
google-auth==2.11.0
google-cloud-core==2.3.2
google-cloud-firestore==2.6.1
googleapis-common-protos==1.56.4
grpcio==1.47.0
grpcio-status==1.47.0
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
proto-plus==1.22.0
protobuf==4.21.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.7.2
requests==2.21.0
rsa==4.9
six==1.16.0
soupsieve==1.7.2
urllib3==1.24.1
Werkzeug==0.14.1
PyJWT~=2.4.0
setuptools~=44.0.0
requests-oauthlib==1.3.1
cryptography==38.0.1